requirements.txt
fastapi~=0.112.2
uvicorn
pydantic~=2.8.2
sqlalchemy~=2.0.34
python-multipart~=0.0.9
passlib[bcrypt]~=1.7.4
python-jose[cryptography]
pillow~=10.4.0
psycopg2-binary
psycopg2~=2.9.9
starlette~=0.38.4
boto3~=1.35.40
python-magic~=0.4.27
loguru~=0.7.2
cffi~=1.17.1
typing_extensions~=4.12.2
pip~=22.2.1
cryptography~=43.0.1
rsa~=4.9
pyasn1~=0.6.1
setuptools~=63.2.0
ecdsa~=0.19.0
botocore~=1.35.40
s3transfer~=0.10.3
jmespath~=1.0.1
six~=1.16.0
pytest~=8.3.2
urllib3~=2.2.3
python-dateutil~=2.9.0.post0
pycparser~=2.22
packaging~=24.1
Jinja2~=3.1.4
bcrypt~=4.0.1
python-dotenv